fix(RegistryV2): fix after review
