+++ back/src/registryV2/typeDefs/private/registryV2.mutations.graphql
@@ -1,0 +1,16 @@
+type Mutation {
+  """
+  USAGE INTERNE
+  Annule une déclaration du registre
+  """
+  cancelRegistryV2Line(
+    "Identifiant unique de la déclaration"
+    publicId: String!
+    "SIRET de l'établissement pour lequel on veut supprimer la déclaration"
+    siret: String!
+    "SIRET de l'établissement délégataire via lequel on veut supprimer la déclaration"
+    delegateSiret: String
+    "Type de registre auquel appartient la déclaration"
+    type: RegistryImportType!
+  ): CancelRegistryV2LineResponse!
+}
