+++ back/src/registryV2/typeDefs/private/registryV2.objects.graphql
@@ -22,3 +22,7 @@ type ChangeAggregateReportAs {
   name: String!
   siret: String
 }
+
+type CancelRegistryV2LineResponse {
+  publicId: ID!
+}
