+++ back/src/registryV2/typeDefs/registryV2.mutations.graphql
@@ -85,16 +85,4 @@ type Mutation {
     "Filtres additionnels sur les dates d'expédition / réception, le code déchet, l'opération de traitement, etc "
     where: RegistryV2ExportWhere
   ): RegistryV2Export!
-
-  "Annule une déclaration du registre"
-  cancelRegistryV2Line(
-    "Identifiant unique de la déclaration"
-    publicId: String!
-    "SIRET de l'établissement pour lequel on veut supprimer la déclaration"
-    siret: String!
-    "SIRET de l'établissement délégataire via lequel on veut supprimer la déclaration"
-    delegateSiret: String
-    "Type de registre auquel appartient la déclaration"
-    type: RegistryImportType!
-  ): CancelRegistryV2LineResponse!
 }
